+++ src/UnitTestEx/AspNetCore/HttpTesterBaseT2.cs
@@ -28,7 +28,7 @@ namespace UnitTestEx.AspNetCore
         /// <param name="testServer">The <see cref="TestServer"/>.</param>
         internal HttpTesterBase(TesterBase owner, TestServer testServer) : base(owner, testServer)
         {
-            _exceptionSuccessExpectations = new ExceptionSuccessExpectations(Owner);
+            _exceptionSuccessExpectations = new ExceptionSuccessExpectations(Implementor);
             _httpResponseExpectations = new HttpResponseExpectations(Owner);
             _responseValueExpectations = new ResponseValueExpectations<TSelf, TValue>(Owner, (TSelf)this);
             _eventExpectations = new EventExpectations(Owner);