+++ tests/UnitTestEx.NUnit.Test/Other/ValidationTest.cs
@@ -1,4 +1,6 @@
-﻿using CoreEx.Validation;
+﻿using CoreEx;
+using CoreEx.Entities;
+using CoreEx.Validation;
 using NUnit.Framework;
 using UnitTestEx.Api.Models;
 using UnitTestEx.Expectations;
@@ -50,6 +52,24 @@ namespace UnitTestEx.NUnit.Test.Other
                     "Identifier is required.",
                     "First Name is required.")
                 .Run<IValidator<Person>, Person>(new Person());
+        }
+
+        [Test]
+        public void Validate_Code_Success()
+        {
+            using var test = ValidationTester.Create();
+
+            test.ExpectSuccess()
+                .RunCode(() => { });
+        }
+
+        [Test]
+        public void Validate_Code_Error()
+        {
+            using var test = ValidationTester.Create();
+
+            test.ExpectErrors("Identifier is required.")
+                .RunCode(() => { throw new ValidationException(new MessageItem[] { MessageItem.CreateErrorMessage("Id", "Identifier is required.") }); });
         }
     }